refactor(client/notes): reorganize note card into modular components and improve formatting
Modularize note card architecture and improve formatting:
  - Extract monolithic note-card.tsx into specialized subcomponents
  - Create dedicated note-card directory with component hierarchy
  - Implement NoteCardHeader for title and metadata display
  - Add NoteCardContent for body and tag rendering
  - Create NoteCardActionsMenu for dropdown functionality
  - Update notes-list to reference new modular card structure
  - Remove original note-card.tsx from notes-dashboard directory

Enhance note card content display and navigation:
  - Add direct navigation links to note details and edit pages
  - Display first line of note content instead of full content for cleaner cards
  - Improve timestamps display with proper formatting
  - Show relative time for note creation date ("2 hours ago" vs exact date)
  - Maintain color-coded card borders for visual organization
