+++ apps/client/src/features/note/components/note-card/index.ts
@@ -1,0 +1,1 @@
+export * from "./note-card";
