+++ apps/client/src/features/note/components/notes-dashboard/index.ts
@@ -1,4 +1,3 @@
-export * from "./note-card";
 export * from "./notes-dashboard-skeleton";
 export * from "./notes-list";
 export * from "./notes-search-filter";
